+++ private/gen/proto/api/buf/alpha/registry/v1alpha1/registryv1alpha1api/repository_commit.pb.go
@@ -53,9 +53,9 @@ type RepositoryCommitService interface {
 		repositoryName string,
 		reference string,
 	) (repositoryCommit *v1alpha1.RepositoryCommit, err error)
-	// GetRepositoryCommitBySequenceID returns the repository commit matching
+	// GetRepositoryCommitBySequenceId returns the repository commit matching
 	// the provided sequence ID and branch, if it exists.
-	GetRepositoryCommitBySequenceID(
+	GetRepositoryCommitBySequenceId(
 		ctx context.Context,
 		repositoryOwner string,
 		repositoryName string,